moved mock data to mock file
